# src/app/api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dojo API",
    "description": "REST API for Dojo LMS — courses, chapters, purchases, promo codes, blog posts, and more. Uses OAuth2 with JWT access tokens and refresh token rotation.",
    "version": "0.1.0",
    "contact": {
      "name": "Dojo API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /api/oauth/token"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "/api/oauth/token",
            "scopes": {
              "read": "Read access to resources",
              "write": "Write access to resources",
              "admin": "Admin access"
            }
          },
          "clientCredentials": {
            "tokenUrl": "/api/oauth/token",
            "scopes": {
              "read": "Read access to resources",
              "write": "Write access to resources",
              "admin": "Admin access"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string", "example": "TypeScript" },
          "slug": { "type": "string", "example": "typescript" }
        }
      },
...
      "YouTubeVideo": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string", "example": "Course Introduction" },
          "videoId": { "type": "string", "example": "dQw4w9WgXcQ" },
          "description": { "type": "string", "nullable": true },
          "position": { "type": "integer" },
          "isPublished": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "FileUploadResponse": {
        "type": "object",
        "properties": {
          "url": { "type": "string", "example": "/api/uploads/550e8400.pdf" },
          "name": { "type": "string", "example": "document.pdf" },
          "size": { "type": "integer", "example": 1048576 },
          "type": { "type": "string", "example": "application/pdf" }
        }
      }
    }
  },
  "paths": {
...
    "/api/tags": {
      "get": {
        "tags": ["Tags"],
        "operationId": "listTags",
        "summary": "List all tags",
        "responses": {
          "200": {
            "description": "Tag list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Tag" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tags"],
        "operationId": "createTag",
        "summary": "Create a tag",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "TypeScript"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tag created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Tag" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "409": { "description": "Already exists" }
        }
      }
    },
...
    "/api/youtube-videos": {
      "get": {
        "tags": ["Videos"],
        "operationId": "listVideos",
        "summary": "List published YouTube videos",
        "responses": {
          "200": {
            "description": "Video list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/YouTubeVideo" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Videos"],
        "operationId": "createVideo",
        "summary": "Add a YouTube video",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "videoId"],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Course Introduction"
                  },
                  "videoId": { "type": "string", "example": "dQw4w9WgXcQ" },
                  "description": { "type": "string" },
                  "position": { "type": "integer", "default": 0 },
                  "isPublished": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Video created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/YouTubeVideo" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/upload": {
      "post": {
        "tags": ["Upload"],
        "operationId": "uploadFile",
        "summary": "Upload a file",
        "description": "Admin only. Max 50 MB. Allowed types: images (JPEG, PNG, GIF, WebP, SVG), PDF, ZIP, text, markdown, JSON, Office docs (DOCX, XLSX, PPTX). Files are stored with UUID filenames.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File uploaded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FileUploadResponse" }
              }
            }
          },
          "400": { "description": "No file, too large, or type not allowed" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden — admin only" }
        }
      }
    },
    "/api/uploads/{path}": {
      "get": {
        "tags": ["Upload"],
        "operationId": "serveFile",
        "summary": "Serve an uploaded file",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "File path (e.g. abc123.pdf)"
          }
        ],
        "responses": {
          "200": {
            "description": "File content with appropriate Content-Type"
          },
          "400": { "description": "Invalid path" },
          "404": { "description": "File not found" }
        }
      }
    }
  }
}